etl/modules/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etl"
version = "4.1.0"
description = "A package of etl modules"
authors = [{ name = "Hassan Idman"}]
requires-python = ">=3.10"

dependencies = [
    "requests==2.32.3",
    "pyspark==3.5.2",
    "pyyaml==6.0.2",  
    "groq==0.12.0",
    "instructor==1.6.4",
    "confluent-kafka==2.1.0",
    "fastavro==1.10.0",
    "apache-airflow-providers-apache-spark==5.0.0"
]
